Delete with two children working.
Also, removed parent as a field. Parent as a field was immutable; hence,
could not cope with a changing data structure, as it kind of pointed off
into history. While this would be interesting for some kind of
time-dependent data structure study, it (1) did not work the way a
mutable parent pointer should, and (2) almost certainly leaked memory,
since the GC shouldn't be able to collect something that has a live
reference.

An attempt was made to use the & pointer type for parent, however, this
inferred into something that was unacceptable for the compiler's region
analysis. So that avenue is closed off until I understand the region
pointer system better, particularly in data structures.

Note that successor is now removed. Successor's implementation relied on
parent being correct. Further, the point of successor was to allow
delete to work correctly. A careful examination of successor will
demonstrate that only half of it is needed for the two-child case in
delete.